Unify different Django-command implementations in scraper
Actually we already had (even before the last commit) a Django
command in scraper.py -- just (a) it didn't do anything without
arguments and didn't print help either, and (b) we also had an
"if __name__ == __main__" block which was disused.

Unify the two implementations into one which is more discoverable.