Tweak diff algorithm to more eagerly absorb small equalities
This means that if e.g. a one-character change is separated by
a two- or three-character equality from a large change, they will
all be combined; which is especially helpful when it cascades to
join two large changes which were separated by multiple small
equalities and smaller changes.

(This is a sort of cherry-pick onto the Python diff-match-patch
of commit bdab645's change to the Javascript implementation.)